// src/enums/ImageRatio.ts
/*
 * @author BSG <dev@bsgroup.eu>
 * @copyright Better Software Group S.A.
 * @version: 1.0
 */

export enum ImageRatio {
  ONE_ONE = "1:1",
  SIXTEEN_NINE = "16:9",
  SIXTEEN_SEVEN = "16:7",
  SIXTEEN_SIX = "16:6",
  SIXTEEN_FIVE = "16:5",
  FOUR_THREE = "4:3",
  THREE_TWO = "3:2",
  EIGHT_FIVE = "8:5",
  TWENTY_TWENTYTHREE = "20:23",
}